+++ src/utils/jwt.ts
@@ -1,29 +1,31 @@
 import { DecodedJwt, EncodedJwt } from '../types/jwt';
+import { parseJson } from './json';
 
-export const decode = (encodedJwt: EncodedJwt): DecodedJwt => {
-  if (!encodedJwt.header || !encodedJwt.payload || !encodedJwt.signature) {
-    throw new Error('Invalid JWT: Incorrect number of segments');
+export const decodeJwt = (encodedJwt: EncodedJwt): DecodedJwt | null => {
+  if (!encodedJwt) {
+    return null;
   }
 
   const decoded = [encodedJwt.header, encodedJwt.payload].map(segment =>
     window.atob(segment),
   );
+  const [header, payload] = decoded;
+  const decodedJwt: DecodedJwt = {
+    header: parseJson(header),
+    payload: parseJson(payload),
+    signature: encodedJwt.signature,
+  };
 
-  try {
-    const [header, payload] = decoded;
-    const decodedJwt: DecodedJwt = {
-      header: JSON.parse(header),
-      payload: JSON.parse(payload),
-      signature: encodedJwt.signature,
-    };
-    return decodedJwt;
-  } catch {
-    throw new Error('Invalid JWT: Payload invalid');
+  if (!decodedJwt.header || !decodedJwt.payload || !decodedJwt.signature) {
+    return null;
   }
+
+  return decodedJwt;
 };
 
-export const split = (jwt: string): EncodedJwt => {
+export const splitJwt = (jwt: string): EncodedJwt => {
   const jwtSegments = jwt.split('.');
+  if (jwtSegments.length !== 3) return null;
   const [header, payload, signature] = jwtSegments;
   return {
     header,